Struct crossbeam_skiplist::SkipSet

pub struct SkipSet<T> { /* fields omitted */ }

A set based on a lock-free skip list.

Implementations

impl<T> SkipSet<T>

pub fn new() -> SkipSet<T>

Returns a new, empty set.

pub fn is_empty(&self) -> bool

Returns true if the set is empty.

pub fn len(&self) -> usize

Returns the number of entries in the set.

If the set is being concurrently modified, consider the returned number just an approximation without any guarantees.

impl<T> SkipSet<T> where
T: Ord,

pub fn front(&self) -> Option<Entry<'_, T>>

Returns the entry with the smallest key.

pub fn back(&self) -> Option<Entry<'_, T>>

Returns the entry with the largest key.

pub fn contains<Q: ?Sized>(&self, key: &Q) -> bool where
T: Borrow<Q>,
Q: Ord,

Returns true if the set contains a value for the specified key.

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<Entry<'_, T>> where
T: Borrow<Q>,
Q: Ord,

Returns an entry with the specified key.

pub fn lower_bound<'a, Q: ?Sized>(
&'a self,
bound: Bound<&Q>
) -> Option<Entry<'a, T>> where
T: Borrow<Q>,
Q: Ord,

Returns an Entry pointing to the lowest element whose key is above the given bound. If no such element is found then None is returned.

pub fn upper_bound<'a, Q: ?Sized>(
&'a self,
bound: Bound<&Q>
) -> Option<Entry<'a, T>> where
T: Borrow<Q>,
Q: Ord,

Returns an Entry pointing to the highest element whose key is below the given bound. If no such element is found then None is returned.

pub fn get_or_insert(&self, key: T) -> Entry<'_, T>

Finds an entry with the specified key, or inserts a new key-value pair if none exist.

pub fn iter(&self) -> Iter<'_, T>

Returns an iterator over all entries in the map.

pub fn range<Q: ?Sized, R>(&self, range: R) -> Range<'_, Q, R, T> where
T: Borrow<Q>,
R: RangeBounds<Q>,
Q: Ord,

Returns an iterator over a subset of entries in the skip list.

impl<T> SkipSet<T> where
T: Ord + Send + 'static,

pub fn insert(&self, key: T) -> Entry<'_, T>

Inserts a key-value pair into the set and returns the new entry.

If there is an existing entry with this key, it will be removed before inserting the new one.

pub fn remove<Q: ?Sized>(&self, key: &Q) -> Option<Entry<'_, T>> where
T: Borrow<Q>,
Q: Ord,

Removes an entry with the specified key from the set and returns it.

pub fn pop_front(&self) -> Option<Entry<'_, T>>

Removes an entry from the front of the map.

pub fn pop_back(&self) -> Option<Entry<'_, T>>

Removes an entry from the back of the map.

pub fn clear(&self)

Iterates over the set and removes every entry.