Struct crossbeam_skiplist::base::SkipList

pub struct SkipList<K, V> { /* fields omitted */ }

A lock-free skip list.

Implementations

impl<K, V> SkipList<K, V>

pub fn new(collector: Collector) -> SkipList<K, V>

Returns a new, empty skip list.

pub fn is_empty(&self) -> bool

Returns true if the skip list is empty.

pub fn len(&self) -> usize

Returns the number of entries in the skip list.

If the skip list is being concurrently modified, consider the returned number just an approximation without any guarantees.

impl<K, V> SkipList<K, V> where
K: Ord,

pub fn front<'a: 'g, 'g>(
&'a self,
guard: &'g Guard
) -> Option<Entry<'a, 'g, K, V>>

Returns the entry with the smallest key.

pub fn back<'a: 'g, 'g>(
&'a self,
guard: &'g Guard
) -> Option<Entry<'a, 'g, K, V>>

Returns the entry with the largest key.

pub fn contains_key<Q: ?Sized>(&self, key: &Q, guard: &Guard) -> bool where
K: Borrow<Q>,
Q: Ord,

Returns true if the map contains a value for the specified key.

pub fn get<'a: 'g, 'g, Q: ?Sized>(
&'a self,
key: &Q,
guard: &'g Guard
) -> Option<Entry<'a, 'g, K, V>> where
K: Borrow<Q>,
Q: Ord,

Returns an entry with the specified key.

pub fn lower_bound<'a: 'g, 'g, Q: ?Sized>(
&'a self,
bound: Bound<&Q>,
guard: &'g Guard
) -> Option<Entry<'a, 'g, K, V>> where
K: Borrow<Q>,
Q: Ord,

Returns an Entry pointing to the lowest element whose key is above the given bound. If no such element is found then None is returned.

pub fn upper_bound<'a: 'g, 'g, Q: ?Sized>(
&'a self,
bound: Bound<&Q>,
guard: &'g Guard
) -> Option<Entry<'a, 'g, K, V>> where
K: Borrow<Q>,
Q: Ord,

Returns an Entry pointing to the highest element whose key is below the given bound. If no such element is found then None is returned.

pub fn get_or_insert(
&self,
key: K,
value: V,
guard: &Guard
) -> RefEntry<'_, K, V>

Finds an entry with the specified key, or inserts a new key-value pair if none exist.

pub fn iter<'a: 'g, 'g>(&'a self, guard: &'g Guard) -> Iter<'a, 'g, K, V>

Returns an iterator over all entries in the skip list.

pub fn ref_iter(&self) -> RefIter<'_, K, V>

Returns an iterator over all entries in the skip list.

pub fn range<'a: 'g, 'g, Q: ?Sized, R>(
&'a self,
range: R,
guard: &'g Guard
) -> Range<'a, 'g, Q, R, K, V> where
K: Borrow<Q>,
R: RangeBounds<Q>,
Q: Ord,

Returns an iterator over a subset of entries in the skip list.

pub fn ref_range<'a, Q: ?Sized, R>(
&'a self,
range: R
) -> RefRange<'a, Q, R, K, V> where
K: Borrow<Q>,
R: RangeBounds<Q>,
Q: Ord,

Returns an iterator over a subset of entries in the skip list.

impl<K, V> SkipList<K, V> where
K: Ord + Send + 'static,
V: Send + 'static,

pub fn insert(&self, key: K, value: V, guard: &Guard) -> RefEntry<'_, K, V>

Inserts a key-value pair into the skip list and returns the new entry.

If there is an existing entry with this key, it will be removed before inserting the new one.

pub fn remove<Q: ?Sized>(
&self,
key: &Q,
guard: &Guard
) -> Option<RefEntry<'_, K, V>> where
K: Borrow<Q>,
Q: Ord,

Removes an entry with the specified key from the map and returns it.

pub fn pop_front(&self, guard: &Guard) -> Option<RefEntry<'_, K, V>>

Removes an entry from the front of the skip list.

pub fn pop_back(&self, guard: &Guard) -> Option<RefEntry<'_, K, V>>

Removes an entry from the back of the skip list.

pub fn clear(&self, guard: &mut Guard)

Iterates over the map and removes every entry.